Split the ValidationStatus class into 2 classes.
