+++ src/main/java/uk/ac/ebi/subs/validator/data/SingleValidationResult.java
@@ -1,24 +1,30 @@
 package uk.ac.ebi.subs.validator.data;
 
 import lombok.ToString;
+import uk.ac.ebi.subs.validator.data.structures.SingleValidationResultStatus;
+import uk.ac.ebi.subs.validator.data.structures.ValidationAuthor;
 
 /**
   * Validation entity result document to store an entity validation result for a specific rule set
   */
-@ToString(callSuper = true)
-public class SingleValidationResult extends AbstractValidationResult {
+@ToString
+public class SingleValidationResult {
 
     private String validationResultUUID;
 
     private ValidationAuthor validationAuthor;
 
-    public SingleValidationResult() {
-    }
+    private SingleValidationResultStatus validationStatus = SingleValidationResultStatus.Pending;
+
+    private String message;
+
+    private String entityUuid;
+
+    public SingleValidationResult() {}
 
     public SingleValidationResult(ValidationAuthor validationAuthor, String entityUuid) {
         this.validationAuthor = validationAuthor;
-        this.setEntityUuid(entityUuid);
-        this.setValidationStatus(ValidationStatus.Pending);
+        this.entityUuid = entityUuid;
     }
 
     public ValidationAuthor getValidationAuthor() {
